+++ src/LevelDBStore/Plugins/Storage/LevelDBStore.cs
@@ -1,6 +1,6 @@
 // Copyright (C) 2015-2024 The Neo Project.
 //
-// LevelDBStore.cs file belongs to the neo project and is free
+// Store.cs file belongs to the neo project and is free
 // software distributed under the MIT software license, see the
 // accompanying file LICENSE in the main directory of the
 // repository or http://www.opensource.org/licenses/mit-license.php
@@ -11,25 +11,57 @@
 
 using Neo.IO.Data.LevelDB;
 using Neo.Persistence;
-using System;
-using System.Linq;
+using System.Collections.Generic;
 
 namespace Neo.Plugins.Storage
 {
-    public class LevelDBStore : Plugin, IStoreProvider
+    internal class LevelDBStore : IStore
     {
-        public override string Description => "Uses LevelDB to store the blockchain data";
+        private readonly DB db;
 
-        public LevelDBStore()
+        public LevelDBStore(string path)
         {
-            StoreFactory.RegisterProvider(this);
+            this.db = DB.Open(path, new Options { CreateIfMissing = true, FilterPolicy = Native.leveldb_filterpolicy_create_bloom(15) });
         }
 
-        public IStore GetStore(string path)
+        public void Delete(byte[] key)
         {
-            if (Environment.CommandLine.Split(' ').Any(p => p == "/repair" || p == "--repair"))
-                DB.Repair(path, Options.Default);
-            return new Store(path);
+            db.Delete(WriteOptions.Default, key);
+        }
+
+        public void Dispose()
+        {
+            db.Dispose();
+        }
+
+        public IEnumerable<(byte[], byte[])> Seek(byte[] prefix, SeekDirection direction = SeekDirection.Forward)
+        {
+            return db.Seek(ReadOptions.Default, prefix, direction, (k, v) => (k, v));
+        }
+
+        public ISnapshot GetSnapshot()
+        {
+            return new Snapshot(db);
+        }
+
+        public void Put(byte[] key, byte[] value)
+        {
+            db.Put(WriteOptions.Default, key, value);
+        }
+
+        public void PutSync(byte[] key, byte[] value)
+        {
+            db.Put(WriteOptions.SyncWrite, key, value);
+        }
+
+        public bool Contains(byte[] key)
+        {
+            return db.Contains(ReadOptions.Default, key);
+        }
+
+        public byte[] TryGet(byte[] key)
+        {
+            return db.Get(ReadOptions.Default, key);
         }
     }
 }
